fix(repository-info): move this util from api to website only
fixes https://github.com/podman-desktop/podman-desktop/issues/16194

breaking changes: remove it from extension api
added by https://github.com/podman-desktop/podman-desktop/pull/13363 (but no related issue)
I don't see usage at https://github.com/search?q=RepositoryInfoParser+language%3ATypeScript&type=code&l=TypeScript&p=1

it's preventing to import things in packages/renderer side
related to https://github.com/podman-desktop/podman-desktop/issues/15496
