+++ website/package.json
@@ -35,6 +35,7 @@
     "clsx": "^2.1.1",
     "docusaurus-plugin-goatcounter": "^4.0.0",
     "docusaurus-plugin-typedoc": "^1.4.2",
+    "hosted-git-info": "^9.0.2",
     "prism-react-renderer": "^2.4.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
