+++ website/src/plugins/github-service.spec.ts
@@ -20,10 +20,9 @@ import { http, HttpResponse } from 'msw';
 import { setupServer } from 'msw/node';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test } from 'vitest';
 
-import { GITHUB_OWNER, GITHUB_REPOSITORY } from '/@api/repository-infos';
-
 import type { GitHubMetadata } from './github-metadata';
 import { GitHubService } from './github-service';
+import { GITHUB_OWNER, GITHUB_REPOSITORY } from './repository-info/repository-infos';
 import { mockReleaseData } from './test/resources/mock-release-data';
 import { mockRepositoryData } from './test/resources/mock-repository-data';
